requirements.txt
click==8.1.6
colorama==0.4.6
gitdb==4.0.10
GitPython==3.1.32
passpy==1.0.2
pyperclip==1.8.2
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
python-gnupg==0.5.1
PyYAML==6.0.1
smmap==5.0.0
validators==0.21.2
